Bot: Refactor messages
