+++ locales/ru/LC_MESSAGES/messages.po
@@ -7,8 +7,8 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PROJECT VERSION\n"
 "Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: 2025-03-30 11:51+0400\n"
-"PO-Revision-Date: 2025-03-30 11:51+0400\n"
+"POT-Creation-Date: 2025-03-30 13:06+0400\n"
+"PO-Revision-Date: 2025-03-30 13:06+0400\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language: ru\n"
 "Language-Team: ru <LL@li.org>\n"
@@ -21,8 +21,24 @@ msgstr ""
 
 #: pars_diary/keyboards.py:18 pars_diary/keyboards.py:57
 msgid "instruction"
-msgstr "Инструкция"
+msgstr "📖 Инструкция"
 
 #: pars_diary/keyboards.py:30
 msgid "start"
-msgstr "Начнём!"
+msgstr "🌟 Начнём!"
+
+#: pars_diary/messages.py:51
+msgid "need auth for action"
+msgstr "🔑 Вам нужно авторизоваться для этого действий."
+
+#: pars_diary/messages.py:90
+msgid "welcome, {first_name}! you need register."
+msgstr "👋 Добро пожаловать, {first_name}!\n\n🔑 Для начала вам нужно пройти быструю регистрацию"
+
+#: pars_diary/messages.py:95
+msgid "1. select_region"
+msgstr "1. 🔎 Выберите свой регион:"
+
+#: pars_diary/messages.py:100
+msgid "2. select cookie"
+msgstr "2. 🍪 Укажите cookie:"
