+++ pars_diary/handlers/about.py
@@ -12,7 +12,7 @@ from aiogram import Router
 from aiogram.filters import Command
 from aiogram.types import Message
 
-from pars_diary.utils.messages import about
+from pars_diary.messages import about
 
 router = Router(name="About bot")
 
@@ -20,4 +20,4 @@ router = Router(name="About bot")
 @router.message(Command("about"))
 async def about_bot(msg: Message) -> None:
     """Рассказывает о проекте."""
-    await msg.answer(about(msg.from_user.language_code))
+    await msg.answer(about())
