+++ pars_diary/messages.py
@@ -1,4 +1,7 @@
-"""Сборщики сообщений."""
+"""Общие сообщения.
+
+Доступные сразу в нескольких обработчиках.
+"""
 
 # TODO @iamlostshe: Добавить поддержку английского языка
 
@@ -8,16 +11,13 @@
 # тг на английском и было бы приятнее получать
 # ответы на английском)
 
-from loguru import logger
+from aiogram.utils.i18n import gettext as _
 
 from pars_diary.config import config
 
 
-def start_old_user(first_name: int | str, lang_code: str | None = "ru") -> str:
+def start_old_user(first_name: str) -> str:
     """Начало для старых пользователей."""
-    if lang_code != "ru":
-        lang_code = "en"
-
     return (
         f"👾 Здравствуйте, {first_name}!\n\n"
         "⭐️ Доступные команды:\n\n"
@@ -33,41 +33,26 @@ def start_old_user(first_name: int | str, lang_code: str | None = "ru") -> str:
     )
 
 
-def error(
-    e: str, lang_code: str | None = "ru", notify: bool | None = False
-) -> str:
+def error_message(e: str) -> str:
     """Сообщение об ошибке."""
-    if lang_code != "ru":
-        lang_code = "en"
-
-    text_e = f"Ошибка во время отправки уведомления ({e})" if notify else e
-
-    logger.error(e)
-
     return (
         "Произошла непредвиденная ошибка, возможно "
         "информация ниже поможет вам понять в чем дело:\n\n"
         "Ошибка:\n\n"
         f"{type(e).__name__}\n\n"
         "Пояснение:\n\n"
-        f"{text_e}\n\n"
+        f"{e}\n\n"
         "Если ошибка произошла не по вашей вине напишите админу @iamlostshe"
     )
 
 
-def not_auth(lang_code: None = "ru") -> str:
+def not_auth() -> str:
     """Если этот контент не доступен без авторизации."""
-    if lang_code != "ru":
-        lang_code = "en"
-
-    return "Для выполнения этого действия вам необходимо зарегистрироваться."
+    return _("need auth for action")
 
 
-def about(lang_code: None = "ru") -> str:
+def about() -> str:
     """Информация о боте."""
-    if lang_code != "ru":
-        lang_code = "en"
-
     return (
         "<b>PARS-DIARY</b> - это проект с открытым исходным кодом, направленный "
         "на улучшение успеваемости учеников путем введения современных технологий.\n\n"
@@ -78,7 +63,7 @@ def about(lang_code: None = "ru") -> str:
         "- <b>Эффективно</b>\n"
         "- <b>Есть функционал уведомлений</b>\n"
         "- <b>Есть функционал умных уведомлений*</b>\n"
-        "- <b>Есть интеграции с нейросетями и инструменты "
+        "- <b>Есть интеграции с AI и инструменты "
         "для быстрого решения домашнего задания</b>\n\n\n"
         "*<b>умные уведомления</b> - [в разработке] уникальная функция "
         "для анализа оценок и простых уведомлений, например:\n\n"
@@ -96,28 +81,20 @@ def about(lang_code: None = "ru") -> str:
     )
 
 
-def registration_0(first_name: int | str, lang_code: str | None = "ru") -> str:
+# Процесс регистрации
+# ===================
+
+
+def registration_0(first_name: str) -> str:
     """Начало для новых пользователей."""
-    if lang_code != "ru":
-        lang_code = "en"
-
-    return (
-        f"👾 Добро пожаловать, {first_name}!\n\n"
-        "Для начала Вам нужно пройти небольшую регистрацию в боте."
-    )
+    return _("welcome, {first_name}! you need register.").format(first_name)
 
 
-def registration_1(lang_code: str | None = "ru") -> str:
+def registration_1() -> str:
     """Начало для новых пользователей."""
-    if lang_code != "ru":
-        lang_code = "en"
-
-    return "1. Укажите Ваш регион:"
+    return _("1. select_region")
 
 
-def registration_2(lang_code: str | None = "ru") -> str:
+def registration_2() -> str:
     """Начало для новых пользователей."""
-    if lang_code != "ru":
-        lang_code = "en"
-
-    return "2. Укажите Ваши cookie:\n\n"
+    return _("2. select cookie")
